Split spec file into two files
The existing spec file was a mess – it was a combination of tests for
both the Rectangle and Square classes. You couldn't independently test
the classes, or see an 'All OK!' message until you got both of them
right.

Which is horrible because they teach you different things: the
`Rectangle` class is meant to introduce you into 'JavaScript class
definition', and the `Square` class is meant to introduce you to `class
inheritance in JavaScript`.

Now you'll be able to get a happy green 'All OK!' message once you've
gotten the `Rectangle` class right, instead of having to wait until you
nail _everything_.